Adding vcat, row-based appending of similar HFrames
vcat works in analogy with hcat, by doing a row-based append of two
similar HFrames. This commit also adds the helper functions schemaEq
which turns the two schemas into maps, and compares them (ie. all
the columns need to have the same names, and the same types to
proceed), and concatBlocks, which concatenates two blocks.